factor out model into a separate file
